tests: Move IBM checks to self-contained testcase
